requirements.txt
pybit>=2.7.0
ccxt>=4.3.90
python-dotenv>=1.0.1
pydantic>=2.7.0
pyyaml>=6.0.2
pandas>=2.2.2
numpy>=1.26.4
prometheus-client>=0.20.0

# Dev
pytest>=8.2.0
pytest-cov>=5.0.0
ruff>=0.6.0
black>=24.8.0
isort>=5.13.2
mypy>=1.10.0
